fix f3read warning message in a certain edge case
The f3read action could warn about corruption rather than warning about misplaced data in the case that the misplaced data had been intended for the first cluster of the chunk currently being examined. This would be a rare case in practice, as any misplacement of data due to fraudulent flash would likely map to a cluster not within the same chunk. Regardless, only the warning message and the count of misplaced clusters were incorrect; no incorrect actions were taken in any case.