don't break the free cluster count when re-marking bad clusters as bad
This commit fixes an edge case that could occur when retesting previously marked-bad clusters. The code would re-mark the clusters as bad and could update the free clusters count in the File System Information sector to an incorrect value.

Also, as of this commit, the most_recently_allocated_data_cluster field of the File System Information sector is no longer updated when marking a cluster bad or free. Updating that field was senseless, as its intended purpose is to amortize linear scanning time when allocating data clusters.