exFAT: don't mistake NoFatChain-allocated clusters for free
A cluster in exFAT is *in use* (or bad) if its corresponding bit in the allocation bitmap is 1, even if its corresponding entry in the FAT is 0. This situation arises whenever virgin clusters have been allocated in NoFatChain mode, as the FAT entries will never have been modified from their virgin state.